Moved server extensions into separate directories
