+++ projects/eclipse/apx_unit_test/.project
@@ -160,19 +160,19 @@
 			<locationURI>PARENT-3-PROJECT_LOC/apx/server/test</locationURI>
 		</link>
 		<link>
-			<name>apx/server_extension/inc</name>
+			<name>apx/server_extension/socket/inc</name>
 			<type>2</type>
-			<location>C:/Users/cogu/repo/c-apx/apx/server_extension/inc</location>
+			<location>C:/Users/cogu/repo/c-apx/apx/server_extension/socket/inc</location>
 		</link>
 		<link>
-			<name>apx/server_extension/src</name>
+			<name>apx/server_extension/socket/src</name>
 			<type>2</type>
-			<location>C:/Users/cogu/repo/c-apx/apx/server_extension/src</location>
+			<location>C:/Users/cogu/repo/c-apx/apx/server_extension/socket/src</location>
 		</link>
 		<link>
-			<name>apx/server_extension/test</name>
+			<name>apx/server_extension/socket/test</name>
 			<type>2</type>
-			<location>C:/Users/cogu/repo/c-apx/apx/server_extension/test</location>
+			<location>C:/Users/cogu/repo/c-apx/apx/server_extension/socket/test</location>
 		</link>
 	</linkedResources>
 </projectDescription>
